Disable zig tests
Not reasonable to depend on Zonation being installed on the platform. Since tutorial results are now included in the package, it is not necessary to run Zonation.

Disable zig tests

Not reasonable to depend on Zonation being installed on the platform. Since tutorial results are now included in the package, it is not necessary to run Zonation.