Parse dat files as is
Previously parameter names from a dat file were cleaned so that they wouldn't include nasty characters (such as " " and "-"). This made it easier to work with them in R, but made it very difficult to write dat files with correct (unmangled) parameter names.

Now parameter names are treated as is.